This is an automated pusher

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -6,12 +6,12 @@
  * Return: the ammount of nodes
 */
 
-
 size_t print_list(const list_t *h)
 {
-size_t count = 0;
-while (h != NULL)
-	{
+	int count = 0;
+
+	while (h != NULL)
+{
 if (h->str != NULL)
 printf("[%u] %s\n", h->len, h->str);
 else
@@ -19,5 +19,5 @@ printf("[0] (nil)\n");
 h = h->next;
 count++;
 }
-return (count);
+	return (count);
 }
